+++ Atlas.Application/CQRS/Orders/Queries/GetLastOrdersPagedListByClient/OrderLookupDto.cs
@@ -77,9 +77,9 @@ namespace Atlas.Application.CQRS.Orders.Queries.GetLastOrdersPagedListByClient
                 .ForMember(dst => dst.IsPickup, opt =>
                     opt.MapFrom(src => src.IsPickup))
                 .ForMember(dst => dst.PurchasePrice, opt =>
-                    opt.MapFrom(src => Math.Ceiling(src.PurchasePrice)))
+                    opt.MapFrom(src => (long)Math.Ceiling(src.PurchasePrice)))
                 .ForMember(dst => dst.SellingPrice, opt =>
-                    opt.MapFrom(src => Math.Ceiling(src.SellingPrice)))
+                    opt.MapFrom(src => (long)Math.Ceiling(src.SellingPrice)))
                 .ForMember(dst => dst.PaymentType, opt =>
                     opt.MapFrom(src => src.PaymentType))
                 .ForMember(dst => dst.IsPrePayed, opt =>
